refactor: 파일 이름 및 경로 수정
