+++ src/App.jsx
@@ -1,11 +1,11 @@
 import { useState } from "react";
 import "./App.css";
 import Header from "./components/Header/Header";
-import RestaurantCategoryFilter from "./components/Restaurant/RestaurantCategoryFilter/RestaurantCategoryFilter";
-import RestaurantList from "./components/Restaurant/RestaurantList/RestaurantList";
-import RestaurantDetailModal from "./components/Modal/RestaurantDetailModal";
-import AddRestaurantModal from "./components/Modal/AddRestaurantModal";
-import getFilteredRestaurant from "./util/getFilteredRestaurant";
+import RestaurantCategoryFilter from "./components/restaurant/RestaurantCategoryFilter/RestaurantCategoryFilter";
+import RestaurantList from "./components/restaurant/RestaurantList/RestaurantList";
+import RestaurantDetailModal from "./components/modal/RestaurantDetailModal";
+import AddRestaurantModal from "./components/modal/AddRestaurantModal";
+import getFilteredRestaurant from "./utils/getFilteredRestaurant";
 
 function App() {
   const [selectedCategory, setCategory] = useState("전체");
